% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanechnikov_l2norm.R
\name{benchmark_epanechnikov_l2norm}
\alias{benchmark_epanechnikov_l2norm}
\title{Benchmark for Epanechnikov L2-Calculations}
\usage{
benchmark_epanechnikov_l2norm(N_seq = 3:12, bw = 0.2)
}
\arguments{
\item{N_seq}{Powers of 2 that we test the implementation for}

\item{bw}{Bandwidth used for estimators}
}
\value{
A data frame containing benchmark information
}
\description{
Benchmark for Epanechnikov L2-Calculations
}
\examples{
bm <- benchmark_epanechnikov_l2norm(N_seq = 3:12)
bm \%>\%
ggplot2::ggplot(ggplot2::aes(x = log(N), y = log(Time), color = Method)) +
ggplot2::geom_line() +
ggplot2::geom_point()
}
