% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{benchmark_kernel_density}
\alias{benchmark_kernel_density}
\title{Benchmark the kernel density evaluation implementations}
\usage{
benchmark_kernel_density(N_seq = 3:12, bw = 0.2)
}
\value{
A data frame containing benchmark information
}
\description{
Benchmark the kernel density evaluation implementations
}
\examples{
bm <- benchmark_kernel_density()
bm \%>\%
bm \%>\%
 ggplot2::ggplot(ggplot2::aes(x = log(N), y = log(Time), color = Method)) +
 ggplot2::geom_line() +
 ggplot2::geom_point() +
 ggplot2::facet_wrap(~Kernel)
}
