% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_density.R
\name{compile_density}
\alias{compile_density}
\title{Compiles kernel density estimate}
\usage{
compile_density(x, bw, kernel = get_kernel("g"))
}
\arguments{
\item{x}{Data used for density estimation}

\item{bw}{Bandwidth}

\item{kernel}{Kernel used for smoothing. Defaults to Gaussian}
}
\value{
The estimated density function
}
\description{
Compiles kernel density estimate
}
\examples{
set.seed(0)
x <- rnorm(1000)
dens_hat <- compile_density(x, 0.1)
ggplot() +
  geom_histogram(aes(x = x, y = after_stat(density))) +
  geom_line(aes(x = x, y = dens_hat(x)), color = "red")
set.seed(NULL)

}
