% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{CompStatKernel}
\alias{CompStatKernel}
\title{Instantiate a CompStatKernel}
\usage{
CompStatKernel(kernel_code)
}
\arguments{
\item{kernel_code}{An alias for the desired kernel.}
}
\value{
The kernel corresponding to the given input.
}
\description{
The valid codes are gaussian/normal, rectangular/uniform, epanechnikov and triangular.
These can also be accessed by the corresponding 1-letter codes.
}
\examples{
set.seed(0)
x <- rnorm(100000)
ekern <- CompStatKernel("e")
profvis::profvis(kernel_density(ekern, x, 0.1, method = "r"))
}
