% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{cvscore.CompStatKernel}
\alias{cvscore.CompStatKernel}
\title{Compute cross-validated out-of-sample loglikelihood for kernel smoother}
\usage{
\method{cvscore}{CompStatKernel}(kernel, x, bw, k = 5, ...)
}
\arguments{
\item{x}{A numeric vector of data points}

\item{bw}{The desired bandwidth}

\item{k}{The number of folds for cross validation}

\item{...}{Additional arguments passed to kernel_density.}

\item{kernel_code}{The abbreviation for the kernel to use for smoothing}
}
\value{
The negative out-of-sample loglikelihood for the kernel smoother
}
\description{
Compute cross-validated out-of-sample loglikelihood for kernel smoother
}
\examples{
x <- rnorm(1000)
bw <- seq(0.01, 1, 0.01)
g <- CompStatKernel("g")
scores <- bw \%>\% purrr::map_dbl(.f = cvscore, kernel = g, x = x)
plot(bw, scores, type = "l")
}
