% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{epa_kernel}
\alias{epa_kernel}
\title{Fast implementation of the Epanechnikov kernel}
\usage{
epa_kernel(x)
}
\arguments{
\item{x}{Vector for the kernel to be evaluated on}
}
\value{
The values of the kernel for the given input
}
\description{
It is possible to make a comparable version of this function where x itself is used to store the values.
This is slightly faster if the kernel is evaluated many time on small inputs, but otherwise slightly slower
}
