% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanechnikov_l2norm.R
\name{epanechnikov_l2norm_binning}
\alias{epanechnikov_l2norm_binning}
\title{L2 Norm based on exact binning}
\usage{
epanechnikov_l2norm_binning(x, r, helper = epanechnikov_l2norm_matrix)
}
\arguments{
\item{x}{A numeric vector}

\item{r}{A positive bandwidth}

\item{helper}{Another epanechnikov-l2norm function to be used within each bin}
}
\value{
The estimated L2-norm of the kernel density estimate for the data
using the epanechnikov kernel.
}
\description{
L2 Norm based on exact binning
}
\examples{
epanechnikov_l2norm_binning(rnorm(1000), 1)
}
