% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanechnikov_l2norm.R
\name{epanechnikov_l2norm_matrix}
\alias{epanechnikov_l2norm_matrix}
\title{L2 Norm through matrix expansion}
\usage{
epanechnikov_l2norm_matrix(x, r)
}
\arguments{
\item{x}{A numeric vector}

\item{r}{A positive bandwidth}
}
\value{
The estimated L2-norm of the kernel density estimate for the data
using the epanechnikov kernel.
}
\description{
L2 Norm through matrix expansion
}
\examples{
epanechnikov_l2norm_matrix(rnorm(100), 1)
}
