% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanechnikov_l2norm.R
\name{epanechnikov_l2norm_naive}
\alias{epanechnikov_l2norm_naive}
\title{L2 Norm for compact data}
\usage{
epanechnikov_l2norm_naive(x, r, do_sort = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{r}{A positive bandwidth}

\item{do_sort}{Whether to sort the input. Defaults to true. Unsorted input
will produce wrong results.}
}
\value{
The estimated L2-norm of the kernel density estimate for the data
using the epanechnikov kernel.
}
\description{
This method only works if the range of the data is shorter than 2 times the
bandwidth. An error is thrown if this in not the case.
}
\examples{
epanechnikov_l2norm_naive(runif(100), 0.51)
}
