% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanechnikov_l2norm.R
\name{epanechnikov_l2norm_running}
\alias{epanechnikov_l2norm_running}
\title{L2 Norm based on running windows}
\usage{
epanechnikov_l2norm_running(x, r)
}
\arguments{
\item{x}{A numeric vector}

\item{r}{A positive bandwidth}
}
\value{
The estimated L2-norm of the kernel density estimate for the data
using the epanechnikov kernel.
}
\description{
Calculates the L2-norm for the kernel density estimate through cleverly
exploiting the sorted data to avoid summing over 0's.
}
\examples{
epanechnikov_l2norm_running(rnorm(1000), 1)
}
