% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{estimate_l2norm}
\alias{estimate_l2norm}
\title{Calculate an estimate of the L2-norm for the current kernel approximation}
\usage{
estimate_l2norm(x, kernel_code, r)
}
\arguments{
\item{x}{The data used for estimation}

\item{kernel_code}{The name of the kernel used}

\item{r}{The bandwidth}
}
\value{
An estimate of the L2-norm for the current kernel approximation
}
\description{
Calculate an estimate of the L2-norm for the current kernel approximation
}
\examples{
estimate_l2norm(1, "e", 1)
estimate_l2norm(rnorm(100), "r", 0.9 * 100^(1/5))
}
