% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglinear_envelope.R
\name{eval_envelope1}
\alias{eval_envelope1}
\title{Evalute a LogLinearEnvelope in a sequence of points}
\usage{
eval_envelope1(x, envelope)
}
\arguments{
\item{x}{The points to evaluate the envelope in.}

\item{envelope}{The envelope.}
}
\value{
The values of envelope(x).
}
\description{
This implementation is particularly slow.
}
\examples{
enve <- LogLinearEnvelope(dnorm, function(z){dnorm(z) * (-z)}, c(-2,0,1))
eval_envelope1(seq(-3,3,0.1), enve)
}
