% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{eval_kdensR}
\alias{eval_kdensR}
\title{Implementation of kernel density evaluation in R.}
\usage{
eval_kdensR(k, grid, x, bw)
}
\arguments{
\item{k}{The kernel evaluation function}

\item{grid}{The grid of values to evaluate the estimated density in}

\item{x}{The data points used to estimate the density}

\item{bw}{The bandwidth}
}
\value{
The estimated density evaluated at the grid points
}
\description{
Implementation of kernel density evaluation in R.
}
