% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_envelopes.R
\name{GaussianEnvelope}
\alias{GaussianEnvelope}
\title{Compute a Gaussian Envelope}
\usage{
GaussianEnvelope(f, rg = c(-3, 3), security = 1 + 1e-09)
}
\arguments{
\item{f}{The function for which to construct the envelope.}

\item{rg}{The range used to compute the fraction of densities.}

\item{security}{Scaling used for alpha.
Safeguards against poor internal optimization.}
}
\value{
A Gaussian envelope of f.
}
\description{
Compute a Gaussian Envelope
}
\examples{
enve <- GaussianEnvelope(dunif, rg=c(0,1), security = 1.1)
}
