% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_density.R
\name{get_kernel}
\alias{get_kernel}
\title{Access commonly used kernels}
\usage{
get_kernel(kernel_code)
}
\arguments{
\item{kernel_code}{An alias for the desired kernel.}
}
\value{
The kernel corresponding to the given input.
}
\description{
The valid codes are gaussian/normal, rectangular/uniform, epanechnikov and triangular.
These can also be accessed by the corresponding 1-letter codes.
}
\examples{
get_kernel("gaussian")
get_kernel("g")
get_kernel("t")
}
