% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_density.R
\name{get_rv}
\alias{get_rv}
\title{Convinience function for encoding common distributions into RV's.}
\usage{
get_rv(key = "n")
}
\arguments{
\item{key}{The name (first letter) of the desired distribution.}
}
\value{
A RandomVariable object corresponding to the distribution specified.
}
\description{
Convinience function for encoding common distributions into RV's.
}
