% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_density.R
\name{ise}
\alias{ise}
\title{Calculate the ISE}
\usage{
ise(d1, d2, lower = -Inf, upper = Inf)
}
\arguments{
\item{d1}{A density function}

\item{d2}{A density function}

\item{lower, upper}{the limits of integration.  Can be infinite.}
}
\value{
A scalar
}
\description{
Calculates Integrated-Square-Error between two densities.
}
\examples{
ise(dnorm, dexp)
ise(dnorm, function(z) dnorm(z, 1))
}
