% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernel_density}
\alias{kernel_density}
\title{Compiles kernel density estimate}
\usage{
kernel_density(
  kernel,
  x,
  bw,
  grid = NULL,
  n = 512,
  from = NULL,
  to = NULL,
  cut = 3,
  method = "c"
)
}
\arguments{
\item{kernel}{Kernel used for smoothing. Defaults to Gaussian}

\item{x}{Data used for density estimation}

\item{bw}{Bandwidth}
}
\value{
The estimated density function
}
\description{
Compiles kernel density estimate
}
\examples{
set.seed(0)
x <- rnorm(1000)
grid <- seq(-3,3,0.01)
kern <- get_kernel("g")
dens_hatR <- density(kern, x, 1, method = "r", grid = grid)
dens_hatC <- density(kern, x, 1, method = "c", grid = grid)
max(abs(dens_hatR - dens_hatC))
ggplot() +
  geom_histogram(aes(x = x, y = after_stat(density))) +
  geom_line(aes(x = z, y = dens_hat), color = "red")
set.seed(NULL)

}
