% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{l2norm.CompStatKernel}
\alias{l2norm.CompStatKernel}
\title{Calculate an estimate of the L2-norm for the current kernel approximation}
\usage{
\method{l2norm}{CompStatKernel}(kernel, x, r, method = "default")
}
\arguments{
\item{x}{The data used for estimation}

\item{r}{The bandwidth}

\item{method}{A string. Specifies the method used to calculate the L2-Norm.}

\item{kernel_code}{The name of the kernel used}
}
\value{
An estimate of the L2-norm for the current kernel approximation
}
\description{
Calculate an estimate of the L2-norm for the current kernel approximation
}
\examples{
set.seed(0)
x <- rnorm(100)
H <- get_kernel("g")
l2norm(H, x, 1)
set.seed(NULL)
}
