% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_bw_constant.R
\name{l2norm}
\alias{l2norm}
\title{Numerical L2-Norm calculation}
\usage{
l2norm(f, lower = -Inf, upper = Inf, ...)
}
\arguments{
\item{f}{an \R function taking a numeric first argument and returning
    a numeric vector of the same length.  Returning a non-finite element will
    generate an error.}

\item{lower, upper}{the limits of integration.  Can be infinite.}

\item{...}{additional arguments passed to integrate or f}
}
\value{
The L2-Norm of f.
}
\description{
Numerical L2-Norm calculation
}
\examples{
l2norm(dnorm)
l2norm(dexp)
l2norm(function(x) 1/sqrt(x), lower = 1, upper = exp(1))
}
