% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_envelopes.R
\name{LaplaceEnvelope}
\alias{LaplaceEnvelope}
\title{Laplacian Envelope Generator}
\usage{
LaplaceEnvelope(f, rg = c(-3, 3), security = 1 + 1e-09, sim_method = 1)
}
\arguments{
\item{f}{The function for which to construct the envelope.}

\item{rg}{The range used to compute the fraction of densities.}

\item{security}{Scaling used for alpha.
Safeguards against poor internal optimization.}
}
\value{
A list of type 'Envelope'
}
\description{
Laplacian Envelope Generator
}
\examples{
enve <- LaplaceEnvelope(dnorm)
}
