% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglinear_envelope.R
\name{LogLinearEnvelope}
\alias{LogLinearEnvelope}
\title{Envelopes for log-concave distributions}
\usage{
LogLinearEnvelope(f, f_prime, tangent_points)
}
\arguments{
\item{f}{A log-concave function.}

\item{f_prime}{The derivative of f.}

\item{tangent_points}{The points at which the envelope should be tangent to f.}
}
\value{
A LogLinearEnvelope object.
}
\description{
Envelopes for log-concave distributions
}
\examples{
LogLinearEnvelope(dnorm, function(z){dnorm(z) * (-z)}, c(-2,0,1))
}
