% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_bw_constant.R
\name{moment}
\alias{moment}
\title{Moments given a density via numerical integration}
\usage{
moment(k, n = 2)
}
\arguments{
\item{k}{A univariate density function.}

\item{n}{A vector of exponents.}
}
\value{
A vector of moments corresponding to the input.
}
\description{
Moments given a density via numerical integration
}
\examples{
moment(dnorm, c(1,2,3,4))
moment(dexp, c(1,2))
moment(dunif, c(1,2))
}
