% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{plot.CompStatKernel}
\alias{plot.CompStatKernel}
\title{Plotting function for CompStatKernel's}
\usage{
\method{plot}{CompStatKernel}(kernel, x = seq(-2, 2, 0.01))
}
\arguments{
\item{kernel}{A CompStatKernel}

\item{x}{A numeric vector with the points the kernel is evaluated in}
}
\value{
A ggplot showing the kernel
}
\description{
Plotting function for CompStatKernel's
}
\examples{
plot(CompStatKernel("g"))
plot(CompStatKernel("e"))
}
