% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglinear_envelope.R
\name{plot.Envelope}
\alias{plot.Envelope}
\title{Plot a LogLinearEnvelope with the function is is an envelope to}
\usage{
\method{plot}{Envelope}(enve, grid = seq(-6, 6, 0.001), logscale = FALSE)
}
\arguments{
\item{enve}{The LogLinearEnvelope.}

\item{grid}{The grid on which the envelope and function are plotted.}
}
\value{
A ggplot-object.
}
\description{
Plot a LogLinearEnvelope with the function is is an envelope to
}
\examples{
enve <- LogLinearEnvelope(dnorm, function(z){dnorm(z) * (-z)}, c(-3, -1, 1, 3))
plot(enve)
plot(enve, logscale=TRUE)
}
