% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_density.R
\name{plot.IterBwEstimate}
\alias{plot.IterBwEstimate}
\title{Plotting for IterBwEstimate-object produced by iter_bw_est}
\usage{
\method{plot}{IterBwEstimate}(IterBwEst)
}
\arguments{
\item{IterBwEst}{A data frame of results produced by iter_bw_est}
}
\value{
A summary plot for the IterBwEstimate-object
}
\description{
Plotting for IterBwEstimate-object produced by iter_bw_est
}
\examples{
set.seed(0)
res_e <- iter_bw_est(runif(1000), maxiter = 30, kernel = "e", cv_k = 3)
res_g <- iter_bw_est(runif(1000), maxiter = 30, kernel = "g", cv_k = 3)
plot(res_e)
plot(res_g)
set.seed(NULL)
}
