% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot density estimates}
\usage{
plot_density(x, bw, kernel_code = "g", grid_length = 200)
}
\arguments{
\item{x}{Data for which density is computed.}

\item{bw}{The bandwidth. Can be a vector of bandwidths.}

\item{kernel_code}{The identifier for which kernel is to be used.}

\item{grid_length}{The number of evaluation points.}
}
\value{
A ggplot with the density estimates for each bandwidth.
}
\description{
Plot density estimates
}
\examples{
plot_density(rnorm(1000), 0.1)
}
