% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_density.R
\name{poisson_prior_log_f_approx}
\alias{poisson_prior_log_f_approx}
\title{Approximation of the poisson prior density}
\usage{
poisson_prior_log_f_approx(y, maxpow = 4)
}
\arguments{
\item{y}{The points to evaluate the density in.}

\item{maxpow}{The order of approximation. Should be at least 2.
The higher the order, the more accurate (and slower) the approximation.}
}
\value{
The approximate density evaluated at y.
}
\description{
Approximation of the poisson prior density
}
\examples{
poisson_prior_log_f_approx(seq(0.001, 1, 0.001))
}
