% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{print.CompStatKernel}
\alias{print.CompStatKernel}
\title{Printing method for CompStatKernel's}
\usage{
\method{print}{CompStatKernel}(kernel)
}
\arguments{
\item{kernel}{}
}
\value{
Prints a tibble of information on the CompStatKernel
}
\description{
Printing method for CompStatKernel's
}
\examples{
"g" \%>\% CompStatKernel()
"e" \%>\% CompStatKernel()
}
