% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_density.R
\name{print.IterBwEstimate}
\alias{print.IterBwEstimate}
\title{Print method for IterBwEstimate}
\usage{
\method{print}{IterBwEstimate}(x)
}
\arguments{
\item{x}{A result from iter_bw_est}
}
\description{
Print method for IterBwEstimate
}
