% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglinear_envelope.R
\name{qLogLinearEnvelope}
\alias{qLogLinearEnvelope}
\title{Quantile function for LogLinearEnvelope}
\usage{
qLogLinearEnvelope(p, enve)
}
\arguments{
\item{p}{A vector evaluation points, each with 0 < p < 1.}

\item{enve}{A LogLinearEnvelope.}
}
\value{
The quantiles of the envelope corresponding to p.
}
\description{
Quantile function for LogLinearEnvelope
}
\examples{
enve <- LogLinearEnvelope(dnorm, function(z){dnorm(z) * (-z)}, c(-2,0,1))
p <- 1:99 / 100
q <- qLogLinearEnvelope(p, enve)
ggplot(mapping = aes(x = q, y = p)) +
geom_line() +
geom_line(aes(y = pnorm(x)), color = "red")
}
