% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_density.R
\name{RandomVariable}
\alias{RandomVariable}
\title{An object mean to represent a distribution.}
\usage{
RandomVariable(f = NULL, log_f = NULL, f_prime = NULL, log_f_prime = NULL)
}
\arguments{
\item{f}{A density function.}

\item{log_f}{The logarithm of the density function.}

\item{f_prime}{The derivative of the density.}

\item{log_f_prime}{The derivative of the log-density.

Note that the log-versions of the density and the differentiated density can
can be inferred from one another. This is done only if NULL is passed.
Any function that could not be inferred will be replaced with an exception raiser.}
}
\value{
A RandomVariable object.
}
\description{
A RandomVariable object should expose a density and a log-density.
}
\examples{
X <- RandomVariable(dnorm, function(z) log(dnorm(z)), function(z) function(z){dnorm(z) * (-z)})
P1 <- RandomVariable(log_f = log_rejec_dens5,log_f_prime = log_rejec_dens_prime)
P_fail <- RandomVariable(f = dnorm)
}
