% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_density.R
\name{rejec_dens1}
\alias{rejec_dens1}
\title{Density evaluation for rejection sampling}
\usage{
rejec_dens1(y)
}
\arguments{
\item{y}{Evaluation point(s).}
}
\value{
The density evaluated in the given points.
}
\description{
Density evaluation for rejection sampling
}
\examples{
rejec_dens1(0)
}
