% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejection_sampling.R
\name{rejec_dens2}
\alias{rejec_dens2}
\title{Density evaluation for rejection sampling}
\usage{
rejec_dens2(y)
}
\arguments{
\item{y}{Evaluation point(s).}
}
\value{
The density evaluated in the given points.
}
\description{
Density evaluation for rejection sampling
}
\examples{
rejec_dens2(0)
}
