% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejections_sampler.R
\name{rejection_sample_factory}
\alias{rejection_sample_factory}
\title{Compile rejection samplers for LogLinearEnvelopes}
\usage{
rejection_sample_factory(enve, alpha = NULL)
}
\arguments{
\item{enve}{A logLinearEnvelope \eqn{g}.}

\item{alpha}{A scalar value ensuring \eqn{\alpha f \leq g}.}
}
\value{
A rejection sampler for the given envelope.
}
\description{
Currently, something is clearly wrong with this implementation.
WORK IN PROGRESS.
}
\examples{
enve <- LogLinearEnvelope(dnorm, function(z){dnorm(z) * (-z)}, c(-2,0,1))
sampler <- rejection_sample_factory(enve, 1/enve$c)
sim <- sampler(5000)
hist(sim)
qqnorm(sim)
qqline(sim)
environment(sampler)$p
environment(sampler)$credibility
}
