% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejections_sampler.R
\name{rejection_sample_naive}
\alias{rejection_sample_naive}
\title{Slow rejection sampling for LogLinearEnvelopes}
\usage{
rejection_sample_naive(n, enve)
}
\arguments{
\item{n}{The number of simulations.}

\item{enve}{A LogLinearEnvelope.}

\item{alpha}{}
}
\value{
Simulations from the distribution that the envelope is based on.
}
\description{
This method is abysmally slow and was only implemented to provide a baseline.
}
\details{
#' Currently, something is clearly wrong with this implementation.
WORK IN PROGRESS.
}
\examples{
enve <- LogLinearEnvelope(dnorm, function(z){dnorm(z) * (-z)}, c(-2,0,1))
rejection_sampler_naive(10, enve)
}
