% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejections_sampler.R
\name{rejection_sampler_naive}
\alias{rejection_sampler_naive}
\title{Slow rejection sampling for arbitrary envelope}
\usage{
rejection_sampler_naive(n, enve)
}
\arguments{
\item{n}{The number of simulations.}

\item{enve}{An object of class 'Envelope'.}

\item{alpha}{}
}
\value{
Simulations from the distribution that the envelope is based on.
}
\description{
This method is abysmally slow and was only implemented to provide a baseline.
}
\examples{
enve <- LogLinearEnvelope(dnorm, function(z){dnorm(z) * (-z)}, c(-2,0,1))
rejection_sampler_naive(10, enve)
}
