% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglinear_envelope.R
\name{rLogLinearEnvelope}
\alias{rLogLinearEnvelope}
\title{Simulate from a LogLinearEnvelope}
\usage{
rLogLinearEnvelope(n, enve)
}
\arguments{
\item{n}{The number of simulations.}

\item{enve}{A LogLinearEnvelope.}
}
\value{
A vector of the desired length containing simulations from the envelope.
}
\description{
Simulate from a LogLinearEnvelope
}
\examples{
enve <- LogLinearEnvelope(dnorm, function(z){dnorm(z) * (-z)}, c(-2,0,1))
y <- rLogLinearEnvelope(10000, enve)
hist(y)
}
