% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_bw_constant.R
\name{second_derivative}
\alias{second_derivative}
\title{Numerical second derivative}
\usage{
second_derivative(f)
}
\arguments{
\item{f}{The function that is be differentiated.}
}
\value{
The second derivative of f.
}
\description{
Calculates the second derivative of the input function using numDeriv.
}
\examples{
second_derivative(log)
second_derivative(exp)
second_derivative(function(x) x^3 / 6)
}
