% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_bw_constant.R
\name{select_bw_constant}
\alias{select_bw_constant}
\title{Bandwidth selection for kernel smoothing}
\usage{
select_bw_constant(f_0 = dnorm, K = dnorm)
}
\arguments{
\item{f_0}{An intial guess at the unknown density.}

\item{K}{The kernel used for smoothing.}
}
\value{
The estimated optimal bandwidth (a positive number).
}
\description{
Select an optimal bandwidth for kernel smoothing based on the AMISE.
This relies on an initial guess f_0 for the unknown density.
}
\examples{
select_bw_constant(dnorm, dnorm)
select_bw_constant(dnorm, function(x) ifelse(abs(x) < 1, 3/4 * (1-x^2), 0))
}
