% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglinear_envelope.R
\name{update.LogLinearEnvelope}
\alias{update.LogLinearEnvelope}
\title{Re-estimate a LogLinearEnvelope}
\usage{
\method{update}{LogLinearEnvelope}(enve, tangent_points)
}
\arguments{
\item{enve}{The previous envelope.}

\item{tangent_points}{The tangent points the new envelope should be fitted to.}
}
\value{
A LogLinearEnvelope.
}
\description{
Mostly useful for adding additional tangent points.
}
\examples{
enve <- LogLinearEnvelope(get_rv("n"))
plot(enve)
enve <- update.LogLinearEnvelope(enve, c(enve$tangent_points, 0))
plot(enve)
}
