% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernel_density}
\alias{kernel_density}
\title{Compiles kernel density estimate}
\usage{
kernel_density(
  kernel,
  x,
  bw,
  grid = NULL,
  n = 512,
  from = NULL,
  to = NULL,
  cut = 3,
  method = "c",
  return_grid = F
)
}
\arguments{
\item{kernel}{Kernel used for smoothing. Defaults to Gaussian}

\item{x}{Data used for density estimation}

\item{bw}{Bandwidth}
}
\value{
The estimated density function
}
\description{
Compiles kernel density estimate
}
\examples{
set.seed(0)
x <- rnorm(1000)
set.seed(NULL)
grid <- density(x, 0.2, kernel = "g")$x

microbenchmark::microbenchmark(
kernel_density("e", x, 0.2, method="c", grid = grid),
kernel_density("e", x, 0.2, method="r", grid = grid),
density(x, 0.2 * sqrt(CompStatKernel("e")$sigma2), kernel="e")
)

microbenchmark::microbenchmark(
kernel_density("g", x, 0.2, method="c"),
kernel_density("g", x, 0.2, method="r"),
density(x, 0.2, kernel="g")
)


}
