% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglinear_envelope.R
\name{plot.Envelope}
\alias{plot.Envelope}
\title{Plot a LogLinearEnvelope with the function is is an envelope to}
\usage{
\method{plot}{Envelope}(enve, grid = seq(-6, 6, 0.001), logscale = FALSE, compare = FALSE)
}
\arguments{
\item{enve}{The LogLinearEnvelope.}

\item{grid}{The grid on which the envelope and function are plotted.}
}
\value{
A ggplot-object.
}
\description{
Plot a LogLinearEnvelope with the function is is an envelope to
}
\examples{
enve <- LogLinearEnvelope(get_rv(), c(-2,0,1))
plot(enve)
plot(enve, logscale=TRUE)
plot(enve, compare=T)
plot(enve, logscale=T, compare=T)
}
