% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_density.R
\name{plot.RandomVariable}
\alias{plot.RandomVariable}
\title{Density plot for random variables}
\usage{
\method{plot}{RandomVariable}(rv, x = seq(-3, 3, 0.01))
}
\arguments{
\item{rv}{A RandomVariable}

\item{x}{A grid of points for which to plot the density}
}
\value{
A plot of the density of the rv
}
\description{
Density plot for random variables
}
