% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_density.R
\name{poisson_prior_log_f}
\alias{poisson_prior_log_f}
\title{Log-density for the poisson prior (Example Distribution)}
\usage{
poisson_prior_log_f(y)
}
\arguments{
\item{y}{The points at which to evaluate the log-density.}
}
\value{
The log-density within the given points.
}
\description{
Log-density for the poisson prior (Example Distribution)
}
\examples{
poisson_prior_log_f(1)
seq(1e-4, 1, 1e-4) \%>\% plot(., poisson_prior_log_f(.), type="l")
}
